backend/requirements.txt
flask
flask-cors
flask-pymongo
opencv-python
